Fixes SI-6563: don't chain error buffer for constructor arg contexts.
The particular problem here was that we were creating constructor arg
context using the base context that already had information about
the previous error ("not found ident e"). This then meant that even
though the application of default arg was type correct, context already
had errors and they were pushed towards the current context and basically
discarding the default arg application. This then lead to args/params mismatch.

One way of fixing was to remove the update of the current context done
in `withCondConstrTyper`. But then the issue is that we get spurious/weird
errors for things like SI-3649 as errors are put into the wrong parent
context and fail to quickly (for the latter they were put
into non-silent context instead of the silent one).

As it is impossible to deal with all the combinations of the buffers
in base context without introducing a lot of special cases,
we won't chain buffers for constructor arg context (similarly as for silent).